chore: workspace notifications refactor
