+++ web/ce/components/workspace-notifications/index.ts
@@ -1,1 +1,2 @@
 export * from "./notification-card/root";
+export * from "./list-root";
